backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
Pillow==10.2.0
aiofiles==23.2.1
pytest==7.4.0
